Initial Commit
Basically this app is meant to help me convert Q-See videos into a readable format for youtube.  It will also help me rename and move the files.

So far, the only thing that I've managed to accomplish is the setup of the textboxes that will be giving the each channel their room name.

I am also able to parse through the set of videos from the directory and seperate out each directory as well as further splitting up the filename.

Splitting up the file name becomes crucial since that is where mos t of my work is spent.

This will also convert them from the QSEE codec to a more general one as well.  For now I am relying on an external program to do that, which I will be including later.