The btnRenameMove should now be able to successfully move the individual files to the appropriate folders.  Still need to add that functionality, but the paths all now point where they need to.
The list of default folders should now append a "/" to the end of the command.

Next thing I'm going to do is, while it is moving the file, it will pars out the date and time and format it correctly.  So something like this : "CH01 - 7-30-10pm - Thursday"

Also, I need to put in some sort of logic to check to see if the files have already been sorted, and do nothing if it has.