Yay! no longer have to deal with the "avc2avi.exe has stopped unexpectedly" error popup :)
Program now handles that.  It does so by checking to see if the source and destination videos has a greater size difference of 2mb, and if so, notes the file name and continues on.

Comparing the converted and original files, there isn't much difference between the sizes.  We're talking up to about 10 kilobytes of observed change.

So anything over 2mb should be a good indication that something went wrong.

Next, need a way to safely stop the whole thing.

Also would like to be able to interact with the UI while it is doing this.

Also need to have a way to rename the files.

Also need to simplify the UI a bit, some unneeded buttons are there.

Making design decision to not have the output directory customizable, since I really don't want to do that right now.