One can now easily get the file name by using "getFileName(path)".
Split up the moveFileToSortedDirectory.  The "moveOriginalFilesToSortedFolders" function will take care of making sure that the default files are all sorted, and in the proper location.

the "moveFileToSortedDirectory" will actually take care of the moving of the file, as well as the renaming of it.

Next going to focus on the UI a bit, get it better organized.

Thinking that for now, the textboxes will only be used when uploading the files to youtube.