Looks like everything is working now.
The first step on the button will now turn green when it is all done.

Gonna take a look into this to see if we can create a busy dialog of some sort while it chugging away at the files.

Also going to take a closer look at the youtube API thing, see if I can get a reference in Visual Studio.