Took sample code from youtube api example repository, got it working for the most part.   It's uploading files, but according to the video manager on youtube, the videos have a upload percentage of 0%.
Quickly added property that will allow me to get get a list of filepaths that needs to be uplaoded.

For now it is only returning the raw list of converted files, need to add logic that will allow it to filter out the files that have already been uploaded.

Also filled in the youtube button.

Not including the json file, because of security reasons.