Now relying on AppSettings to keep track of what the channel names are as well as the list of videos that have been uploaded.
One should now be able to easily enable / disable the textboxes for naming the channels. (setTextBoxesEnabled)

Each textbox will update the stored settings, all of them will execute the  txt_TextUpdate command

Need to make properties that will return the sorted files, and one that returns the converted files.

In process of getting the lstFilesToUpload to correctly filter out the ones that have already been uploaded.