Youtube uploader should now throw "FileUploadFailure" event when the upload doesn't complete.
getDateTime(path) should now be returning the correct string.

Have stub for handling upload failure on main form

One should now be able to iterate through all the files that have yet to be uploaded.

We are now also keeping track of the video titles that we upload, this should make creating playlists easier later.

TODO:
  Create progress bars on how many files remain to be uploaded, how many files there remains to be converted, and how many files there remains to be sorted.
  Figure out how to create and add videos to a playlist
    See if we can do this after we upload each file.