Added a bunch of properties to the YouTubeUploader that will help with getting informaiton from the path.
Moved the processing of each individual file into the YouTubeUploader class - running into issue where the form would not wait for this to stop, and there isn't a good way to force it to.

Added three more events to hte YouTubeUploader - UpdateFilePath, UpdateOverallStatus, and updateOverallProgress.  The OveralStatus will send back a stirng stating how many items out of the total amount we have uploaded.  The OverallProgress will update the progress bar for the overall upload percentage.

Added event handlers to the form for the three above mentioned events.