Getting the day of the week no longer returns the ".avi" setting
getChannelNumber should no longer be returning a null reference, causing an error on Program.cs.  I forgot to change where it should be grabbing the file path :(

The total number of files uploaded should now appear when the form starts, and correctly update itself now during the uploading of the next file.  It should also now update itself when they clear the upload list.

Default description for the videos are now : "This is a security camera footage from the Madison Terror Trial's Haunted House of (year).  Primary purpose of video survalance is for the safety of the scarers, and we want the ability to see what happened if something did.  As a added side benifit, we can also enjoy the scares themselves at a later point.

At this point I believe that I have the conversion going correctly, and the uploading of the files to Youtube to be okay.  Next is to sort these into playlists.

Theoritically, I should be able to go through the users current videos and assign each one to their appropriate playlist.

More details will be following on how a playlist should be setup.