Basically copied over the sample file that Google uses. It's up and running, currently tied to the new "Create Playlists" button.
Has no customization yet, will be adding that stuff in here shortly

Will be looking into more stuff, basically what needs to happen now is that I need to associate the video with the correct playlist.  Wether this occurs before or after the playlists are created, is to be determined.

Currently if the playlist is already existing, this will create a new one with an identical name.  Need to find a way to prevent that.