Fixed the progress bar, didn't realize that the progress.BytesSent was keeping count of all the bytes sent so far, so removed the "sumOfUploadedBytes" property, and relying on the built in one instead :\ Now getting the total number of bytes differently, just opening up a stream to the file, getting the length, and closing the stream.
Went and hardcoded the year, since I was using DateTime.Now to get the year, which will start to be wrong here in about a half hour for the Central Time Zone . . . Happy New Year :)

Can now control the upload chunk size, currently have it to a about 2mb to allow a somewhat smooth progress bar on the UI.

Seem to be running into a bottleneck somewhere, the videos are taking about 6-12  minutes each to upload fully . . . not quite sure what is going on there
Also it appears that the remaining videos to be uploaded is wrong as well :\

Really should create a class that stores all the video information in it such as the title, tags, location, dates and times, room name, room order and so forth.  And maybe have it as a local variable.